Reorganized tests, combined parameter standalone tests and bson type tests (minmax excepted)
